feat(sumNumbers): 입력값에 있는 실수를 처리하도록 int에서 double로 변경
입력값에 실수도 포함될 수 있기에 int에서 double로 타입을 변경.
이에 따라 StringAdder.add,  Application , inputValidator.validate에서도 double을
처리할 수 있도록 수정.

+++ src/main/java/calculator/Application.java
@@ -6,7 +6,7 @@ public class Application {
         StringAdder stringAdder = new StringAdder();
 
         String input = ConsoleIO.getInput();
-        int result = stringAdder.add(input);
+        double result = stringAdder.add(input);
         ConsoleIO.printResult(result);
     }
 }
